핑 로그 표시 옵션 추가, SetPing 함수 추가
- TcpTransportOption에 핑 로그 표시 옵션 추가, SetPing 함수 추가
- enum 값 비교할 때 int 캐스팅 삭제

+++ funapi-plugin-unity/Assets/Funapi/FunapiTransport.cs
@@ -1053,7 +1053,7 @@ namespace Fun
 
             if (body_length > 0)
             {
-                if ((int)state_ < (int)State.kConnected)
+                if (state_ < State.kConnected)
                 {
                     FunDebug.Log("Unexpected message. state:{0}", state_);
                     return false;
@@ -1187,7 +1187,7 @@ namespace Fun
         {
             get
             {
-                return sock_ != null && sock_.Connected && (int)state_ >= (int)State.kConnected;
+                return sock_ != null && sock_.Connected && state_ >= State.kConnected;
             }
         }
 
@@ -1486,7 +1486,7 @@ namespace Fun
 
         internal override bool Started
         {
-            get { return sock_ != null && (int)state_ >= (int)State.kConnected; }
+            get { return sock_ != null && state_ >= State.kConnected; }
         }
 
         public override bool IsDatagram
@@ -1735,7 +1735,7 @@ namespace Fun
 
         internal override bool Started
         {
-            get { return (int)state_ >= (int)State.kConnected; }
+            get { return state_ >= State.kConnected; }
         }
 
         public override bool IsRequestResponse
